NF: move PreferencesTest and PreferencesSimpleTest to anki.preferences package
