+++ AnkiDroid/src/test/java/com/ichi2/anki/preferences/PreferencesSimpleTest.kt
@@ -13,9 +13,9 @@
  *  You should have received a copy of the GNU General Public License along with
  *  this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.ichi2.anki
+package com.ichi2.anki.preferences
 
-import com.ichi2.anki.preferences.Preferences
+import com.ichi2.anki.RobolectricTest
 import com.ichi2.preferences.HeaderPreference
 import org.hamcrest.MatcherAssert.assertThat
 import org.hamcrest.Matchers.equalTo
